+++ resources/datos_pasajeros.csv
@@ -1,0 +1,2 @@
+originStation,destinationStation,firstName,primerApellido,segundoApellido,dni,email,phone,emailBuyer,phoneBuyer,bankCard,expirationDate,cvv
+VALENCIA JOAQUÍN SOROLLA,BARCELONA-SANTS,John,Doe,Lopez,46131651E,test@qa.com,696824570,comprador@qa.com,696824571,4000 0000 0000 1000,03/30,990
