+++ resources/datos_pasajeros_blank_payment_data_5d.csv
@@ -1,0 +1,2 @@
+﻿originStation,destinationStation,firstName,primerApellido,segundoApellido,dni,email,phone,emailBuyer,phoneBuyer,bankCard,expirationDate,cvv
+VALENCIA JOAQUÍN SOROLLA,BARCELONA-SANTS,John,Doe,Lopez,46131651E,test@qa.com,696824570,comprador@qa.com,696824571,,,
