+++ resources/precios_trayectos.csv
@@ -1,0 +1,2 @@
+﻿HOME_PAGE,SELECCIONA_TU_VIAJE_PAGE,INTRODUCE_TUS_DATOS_PAGE,PERSONALIZA_TU_VIAJE_PAGE,COMPRA_PAGE,PASARELA_PAGO_PAGE
+basicFarePrice,totalPriceTrip,totalPriceData,totalPriceTrip,totalPricePersonalize,totalPriceTrip,totalPriceCompra,totalPriceTrip,totalPricePasarela,totalPriceTrip
