--- src/java/resources/datos_pasajeros_error_15d.csv
@@ -1,2 +1,0 @@
-﻿originStation,destinationStation,firstNameError,primerApellido,segundoApellido,dni,email,phone
-VALENCIA JOAQUÍN SOROLLA,BARCELONA-SANTS,1,Doe,Lopez,46131651E,test@qa.com,696824570
