+++ src/test/java/tests/InvalidCardPaymentTest.java
@@ -1,0 +1,98 @@
+package tests;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.firefox.FirefoxOptions;
+import org.testng.annotations.*;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import pages.*;
+import steps.Steps;
+import utils.CSVDataProvider;
+import utils.DriverManager;
+
+import static pages.BasePage.TIMEOUT;
+
+public class InvalidCardPaymentTest {
+
+    private WebDriver webDriver;
+    private SeleccionarTuViajePage seleccionarTuViajePage;
+    private IntroduceTusDatosPage introduceTusDatosPage;
+    private PersonalizaTuViajePage personalizaTuViajePage;
+    private CompraPage compraPage;
+    private PasarelaPagoPage pasarelaPagoPage;
+    private Steps steps;
+
+    @DataProvider(name = "paymentData")
+    public Object[][] getPaymentData() {
+        return CSVDataProvider.readDatosPasajeros();
+    }
+
+    @DataProvider(name = "routeData")
+    public Object[][] getRouteData() {
+        return CSVDataProvider.readPreciosTrayectos();
+    }
+
+    @BeforeMethod
+    @Parameters("browser")
+    public void setup(@Optional("chrome") String browser) {
+        webDriver = DriverManager.getDriver(browser);
+        webDriver.manage().timeouts().implicitlyWait(TIMEOUT);
+        webDriver.manage().window().maximize();
+        webDriver.get("https://www.renfe.com/es/es");
+
+        seleccionarTuViajePage = new SeleccionarTuViajePage(webDriver);
+        introduceTusDatosPage = new IntroduceTusDatosPage(webDriver);
+        steps = new Steps(webDriver);
+
+        webDriver.manage().timeouts().implicitlyWait(TIMEOUT);
+        webDriver.manage().window().maximize();
+        webDriver.get("https://www.renfe.com/es/es");
+
+        seleccionarTuViajePage = new SeleccionarTuViajePage(webDriver);
+        introduceTusDatosPage = new IntroduceTusDatosPage(webDriver);
+        personalizaTuViajePage = new PersonalizaTuViajePage(webDriver);
+        compraPage = new CompraPage(webDriver);
+        pasarelaPagoPage = new PasarelaPagoPage(webDriver);
+        steps = new Steps(webDriver);
+    }
+
+    @Test(dataProvider = "paymentData")
+    public void InvalidCardPaymentTest(
+            String originStation,
+            String destinationStation,
+            String firstName,
+            String primerApellido,
+            String segundoApellido,
+            String dni,
+            String email,
+            String phone,
+            String emailBuyer,
+            String phoneBuyer,
+            String bankCard,
+            String expirationDate,
+            String cvv) {
+        steps.performSearchOriginAndDestinationStation(originStation, destinationStation);
+        steps.selectDepartureDate();
+        steps.selectTrainAndFare();
+        steps.getAndStoreDynamicPrice();
+        steps.verifyAndConfirmTravel();
+        steps.clickPopUpAndLinkAppear();
+        steps.verifyPriceIsEqualInData();
+        steps.introduceYourDataAndConfirm(firstName, primerApellido, segundoApellido, dni, email, phone);
+        steps.verifyPriceIsEqualInPersonalize();
+        steps.confirmPersonalization();
+        steps.verifyPriceIsEqualInCompra();
+        steps.confirmPaymentData(emailBuyer, phoneBuyer);
+        steps.payment(bankCard, expirationDate, cvv);
+    }
+
+    @AfterMethod
+    public void tearDown() {
+        if (webDriver != null) {
+            webDriver.quit();
+        }
+    }
+}
